requirements.txt
discord.py
tinydb
tinydb_smartcache
python-dotenv
